backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "google-genai>=1.63.0",
    "pymupdf>=1.27.1",
    "python-docx>=1.2.0",
]
